pyproject.toml
[tool.poetry]
name = "k45-simulation"
version = "0.1.0"
description = ""
authors = ["Safarnejad <safarnejad.ho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
Faker = "^22.2.0"
kafka-python = "^2.0.2"
jupyterlab = "^4.0.10"
pyspark = { version = "3.5.0", extras = ["kafka"] }
findspark = "^2.0.1"
py4j = "^0.10.9.7"
pandas = "^2.2.0"
apache-airflow = "^2.8.1"
apache-airflow-providers-postgres = "^5.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
